services/imu/requirements.txt
AHRS==0.2.2
cycler==0.10.0
grpcio==1.28.1
grpcio-tools==1.28.1
kiwisolver==1.2.0
matplotlib==3.2.1
mpu9250-jmdev==1.0.10
numpy==1.18.4
pkg-resources==0.0.0
protobuf==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.4.1
six==1.14.0
smbus==1.1.post2
smbus2==0.3.0
